Trait rc5::rc5_dyn::RC5Algo

pub trait RC5Algo {
    // Required methods
    fn encrypt<'a>(
        &self,
        bytes: &'a mut [u8]
    ) -> Result<&'a mut [u8], RC5AlgoError>;
    fn decrypt<'a>(
        &self,
        bytes: &'a mut [u8]
    ) -> Result<&'a mut [u8], RC5AlgoError>;
    fn block_size(&self) -> usize;
}

The RC5Algo trait provides methods for encrypting and decrypting data using the RC5 block cipher algorithm. This trait is useful when the RC5 algorithm needs to be constructed with a block size that is determined at runtime.

Required Methods

fn encrypt<'a>(&self, bytes: &'a mut [u8]) -> Result<&'a mut [u8], RC5AlgoError>

Encrypts the given slice of bytes in place using the RC5 block cipher algorithm.

Returns a reference to the encrypted bytes on success, or an RC5AlgoError if the encryption failed.

fn decrypt<'a>(&self, bytes: &'a mut [u8]) -> Result<&'a mut [u8], RC5AlgoError>

Decrypts the given slice of bytes in place using the RC5 block cipher algorithm.

Returns a reference to the decrypted bytes on success, or an RC5AlgoError if the decryption failed.

fn block_size(&self) -> usize

Return the block size of the algorithm