Struct rc5::RC5

pub struct RC5<T> { /* private fields */ }

The RC5 struct represents an instance of the RC5 block cipher algorithm.

The RC5 struct provides methods for initializing the algorithm with a key and a number of rounds, as well as methods for encrypting and decrypting individual words and blocks of data.

Implementations

impl<T> RC5<T>where T: Unsigned + FromToLeBytes<T> + FromU64 + WrappingAdd + WrappingSub + Copy + BitXor<T, Output = T> + BitRotabable<T, Output = T> + BitRotabable<u32, Output = T>,

pub fn new(repetitions: u8, key: &[u8]) -> Result<RC5<T>, RC5InitError>

Creates a new RC5 instance with the given key, number of repetitions, and block size.

The key parameter is a slice of bytes representing the secret key used to encrypt and decrypt the data. The repetitions parameter specifies the number of times the key expansion process should be repeated (typically 12, 16, or 20), and the p and q parameters are computed from the base of natura logorithms and golden ration numbers for the specific word size.

Examples

use rc5::RC5;

let key = b"my secret key";
let rc5 = RC5::<u32>::new(12, key);

pub fn encrypt_words(&self, a: &mut T, b: &mut T)

Encrypts the two-word block represented by the references a and b.

The a and b parameters are mutable references to the two words (of type T) to be encrypted. The encrypted values are written back to the same references.

Examples

use rc5::{RC5, RC5InitError};

let key = b"my secret key";
let rc5 = RC5::<u32>::new(12, key)?;

let mut a = 0x12345678;
let mut b = 0x9ABCDEF0;

rc5.encrypt_words(&mut a, &mut b);

assert_eq!(a, 0x4907F9B8);
assert_eq!(b, 0x3F53A579);

pub fn encrypt_word_bytes( &self, a_bytes: &mut [u8; { _ }], b_bytes: &mut [u8; { _ }] )

Encrypts the two-word block represented by the references a_bytes and b_bytes.

The a_bytes and b_bytes parameters are mutable references to the two words (of type [u8; std::mem::size_of::<T>()]) to be encrypted. The encrypted values are written back to the same array references.

Examples

use rc5::{RC5, RC5InitError};

let key = b"my secret key";
let rc5 = RC5::<u32>::new(12, key)?;

let mut a_bytes = [0x78, 0x56, 0x34, 0x12];
let mut b_bytes = [0xF0, 0xDE, 0xBC, 0x9A];

rc5.encrypt_word_bytes(&mut a_bytes, &mut b_bytes);

assert_eq!(a_bytes, [0xB8, 0xF9, 0x07, 0x49]);
assert_eq!(b_bytes, [0x79, 0xA5, 0x53, 0x3F]);

pub fn encrypt_block(&self, bytes: &mut [u8; { _ }])

Encrypts the two-word block represented by the reference bytes.

The bytes parameter is a mutable reference to the two words (of type [u8; std::mem::size_of::<T>() * NUM_WORDS_PER_BLOCK]) to be encrypted. The encrypted values are written back to the same array reference.

Examples

use rc5::{RC5, RC5InitError};

let key = b"my secret key";
let rc5 = RC5::<u32>::new(12, key)?;

let mut bytes = [0x78, 0x56, 0x34, 0x12, 0xF0, 0xDE, 0xBC, 0x9A];

rc5.encrypt_block(&mut bytes);

assert_eq!(bytes, [0xB8, 0xF9, 0x07, 0x49, 0x79, 0xA5, 0x53, 0x3F]);

pub fn decrypt_words(&self, a: &mut T, b: &mut T)

Decrypts the two-word block represented by the references a and b.

The a and b parameters are mutable references to the two words (of type T) to be decrypted. The decrypted values are written back to the same references.

Examples

use rc5::{RC5, RC5InitError};

let key = b"my secret key";
let rc5 = RC5::<u32>::new(12, key)?;

let mut a = 0x4907F9B8;
let mut b = 0x3F53A579;

rc5.decrypt_words(&mut a, &mut b);

assert_eq!(a, 0x12345678);
assert_eq!(b, 0x9ABCDEF0);

pub fn decrypt_word_bytes( &self, a_bytes: &mut [u8; { _ }], b_bytes: &mut [u8; { _ }] )

Decrypts the two-word block represented by the references a_bytes and b_bytes.

The a_bytes and b_bytes parameters are mutable references to the two words (of type [u8; std::mem::size_of::<T>()]) to be decrypted. The decrypted values are written back to the same array references.

Examples

use rc5::{RC5, RC5InitError};

let key = b"my secret key";
let rc5 = RC5::<u32>::new(12, key)?;

let mut a_bytes = [0xB8, 0xF9, 0x07, 0x49];
let mut b_bytes = [0x79, 0xA5, 0x53, 0x3F];

rc5.decrypt_word_bytes(&mut a_bytes, &mut b_bytes);

assert_eq!(a_bytes, [0x78, 0x56, 0x34, 0x12]);
assert_eq!(b_bytes, [0xF0, 0xDE, 0xBC, 0x9A]);

pub fn decrypt_block(&self, bytes: &mut [u8; { _ }])

Decrypts the two-word block represented by the reference bytes.

The bytes parameter is a mutable reference to the two words (of type [u8; std::mem::size_of::<T>() * NUM_WORDS_PER_BLOCK]) to be decrypted. The decrypted values are written back to the same array reference.

Examples

use rc5::{RC5, RC5InitError};

let key = b"my secret key";
let rc5 = RC5::<u32>::new(12, key)?;

let mut bytes = [0xB8, 0xF9, 0x07, 0x49, 0x79, 0xA5, 0x53, 0x3F];

rc5.decrypt_block(&mut bytes);

assert_eq!(bytes, [0x78, 0x56, 0x34, 0x12, 0xF0, 0xDE, 0xBC, 0x9A]);